chore: Update Header component with useUser hook
This commit updates the Header component to use the `useUser` hook from the `@clerk/nextjs` library. This hook allows us to access the user object, which we can use to customize the header based on the user's authentication status. This change improves the functionality and user experience of the header.